Moved checks for background job notification to main activity
