+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/onboarding/OnboardingNotificationsFragment.kt
@@ -59,87 +59,11 @@ class OnboardingNotificationsFragment : Fragment() {
 
     private fun setButtonOnClickListener() {
         binding.onboardingButtonNext.setOnClickListener {
-            checkForBackgroundJobDisabled()
+            navigateToMain()
         }
         binding.onboardingButtonBack.buttonIcon.setOnClickListener {
             (activity as OnboardingActivity).goBack()
         }
-    }
-
-    private fun checkForBackgroundJobDisabled() {
-        if (!ConnectivityHelper.isBackgroundJobEnabled(requireActivity())) {
-            showBackgroundJobDisabledNotification()
-        } else {
-            checkForEnergyOptimizedEnabled()
-        }
-    }
-
-    private fun checkForEnergyOptimizedEnabled() {
-        if (!PowerManagementHelper.isIgnoringBatteryOptimizations(requireActivity())) {
-            showEnergyOptimizedEnabledForBackground()
-        } else {
-            navigateToMain()
-        }
-    }
-
-    private fun showBackgroundJobDisabledNotification() {
-        val dialog = DialogHelper.DialogInstance(
-            requireActivity(),
-            R.string.onboarding_background_fetch_dialog_headline,
-            R.string.onboarding_background_fetch_dialog_body,
-            R.string.onboarding_background_fetch_dialog_button_positive,
-            R.string.onboarding_background_fetch_dialog_button_negative,
-            false, {
-                val intent = Intent(
-                    ACTION_APPLICATION_DETAILS_SETTINGS,
-                    Uri.fromParts("package", requireContext().packageName, null)
-                )
-                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-
-                navigateToMain()
-
-                startActivity(intent)
-                // show battery optimization system dialog after background processing dialog
-                checkForEnergyOptimizedEnabled()
-            }, {
-                // declined, show additional dialog explaining manual risk calculation
-                showManualCheckingRequiredDialog()
-            })
-        DialogHelper.showDialog(dialog)
-    }
-
-    private fun showEnergyOptimizedEnabledForBackground() {
-        val dialog = DialogHelper.DialogInstance(
-            requireActivity(),
-            R.string.onboarding_energy_optimized_dialog_headline,
-            R.string.onboarding_energy_optimized_dialog_body,
-            R.string.onboarding_energy_optimized_dialog_button_positive,
-            R.string.onboarding_energy_optimized_dialog_button_negative,
-            false, {
-                // go to battery optimization
-                ExternalActionHelper.toBatteryOptimizationSettings(requireContext())
-                LocalData.energyOptimizedExplanationDialogWasShown(true)
-                navigateToMain()
-            }, {
-                // keep battery optimization enabled
-                LocalData.energyOptimizedExplanationDialogWasShown(true)
-                showManualCheckingRequiredDialog()
-            })
-        DialogHelper.showDialog(dialog)
-    }
-
-    private fun showManualCheckingRequiredDialog() {
-        val dialog = DialogHelper.DialogInstance(
-            requireActivity(),
-            R.string.onboarding_manual_required_dialog_headline,
-            R.string.onboarding_manual_required_dialog_body,
-            R.string.onboarding_manual_required_dialog_button,
-            null,
-            false, {
-                navigateToMain()
-            }
-        )
-        DialogHelper.showDialog(dialog)
     }
 
     private fun navigateToMain() {
